moved the Text utils
