+++ PayMe.sln
@@ -17,6 +17,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core.UnitTests", "tests\Cor
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain.UnitTests", "tests\Domain.UnitTests\Domain.UnitTests.csproj", "{85C3C3FB-9379-48D8-B63C-A7BDF3D3832C}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Packages", "Packages", "{65E5A51F-5B2B-4516-A7C9-430239AD6689}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Text", "src\Depends\Text\Text.csproj", "{F15462FD-9624-4193-A31C-90C7F2428EEB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -90,6 +94,18 @@ Global
 		{85C3C3FB-9379-48D8-B63C-A7BDF3D3832C}.Release|x64.Build.0 = Release|Any CPU
 		{85C3C3FB-9379-48D8-B63C-A7BDF3D3832C}.Release|x86.ActiveCfg = Release|Any CPU
 		{85C3C3FB-9379-48D8-B63C-A7BDF3D3832C}.Release|x86.Build.0 = Release|Any CPU
+		{F15462FD-9624-4193-A31C-90C7F2428EEB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F15462FD-9624-4193-A31C-90C7F2428EEB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F15462FD-9624-4193-A31C-90C7F2428EEB}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F15462FD-9624-4193-A31C-90C7F2428EEB}.Debug|x64.Build.0 = Debug|Any CPU
+		{F15462FD-9624-4193-A31C-90C7F2428EEB}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F15462FD-9624-4193-A31C-90C7F2428EEB}.Debug|x86.Build.0 = Debug|Any CPU
+		{F15462FD-9624-4193-A31C-90C7F2428EEB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F15462FD-9624-4193-A31C-90C7F2428EEB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F15462FD-9624-4193-A31C-90C7F2428EEB}.Release|x64.ActiveCfg = Release|Any CPU
+		{F15462FD-9624-4193-A31C-90C7F2428EEB}.Release|x64.Build.0 = Release|Any CPU
+		{F15462FD-9624-4193-A31C-90C7F2428EEB}.Release|x86.ActiveCfg = Release|Any CPU
+		{F15462FD-9624-4193-A31C-90C7F2428EEB}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{9E24C4FC-AC52-4F76-8612-A3C2E5F8ED53} = {11D688C8-4A43-4F59-ACA9-BB704B14970E}
@@ -97,5 +113,7 @@ Global
 		{5FE4809E-F4F3-45EF-AC39-09468BA58F4B} = {11D688C8-4A43-4F59-ACA9-BB704B14970E}
 		{FB1CBDE9-49D4-4626-8DFB-167D41A6E87C} = {CBEC906D-A625-48D8-8AD4-335C018812D2}
 		{85C3C3FB-9379-48D8-B63C-A7BDF3D3832C} = {CBEC906D-A625-48D8-8AD4-335C018812D2}
+		{65E5A51F-5B2B-4516-A7C9-430239AD6689} = {11D688C8-4A43-4F59-ACA9-BB704B14970E}
+		{F15462FD-9624-4193-A31C-90C7F2428EEB} = {65E5A51F-5B2B-4516-A7C9-430239AD6689}
 	EndGlobalSection
 EndGlobal
