+++ src/Core/Services/IPaymentService.cs
@@ -1,5 +1,5 @@
 using Core.Entities;
-using Core.Utils;
+using Text;
 
 namespace Core.Services
 {
