+++ src/Depends/Text/Text.csproj
@@ -1,9 +1,5 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-  <ItemGroup>
-    <ProjectReference Include="..\Depends\Text\Text.csproj" />
-  </ItemGroup>
-
   <PropertyGroup>
     <LangVersion>8.0</LangVersion>
     <TargetFramework>netstandard2.1</TargetFramework>
